backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg2-binary
alembic

pydantic
pydantic-settings

python-jose
passlib[bcrypt]

python-multipart
